Fix bug when casting to uint8_t
Using lexical_cast from string to uint8_t doesn't cast the string to a
numeric value, but the ASCII value of the number in the string.

+++ src/house.cpp
@@ -731,7 +731,7 @@ bool Houses::loadHousesXML(const std::string& filename)
 		Position entryPos(
 			pugi::cast<uint16_t>(houseNode.attribute("entryx").value()),
 			pugi::cast<uint16_t>(houseNode.attribute("entryy").value()),
-			pugi::cast<uint8_t>(houseNode.attribute("entryz").value())
+			pugi::cast<uint16_t>(houseNode.attribute("entryz").value())
 		);
 		if (entryPos.x == 0 && entryPos.y == 0 && entryPos.z == 0) {
 			std::cout << "Warning: [Houses::loadHousesXML] House entry not set"
